move SuccessorsSelect to Successors
